refactor: 예외 처리 개선 및 불필요한 코드 제거

+++ modi/rental-service/src/main/java/com/coc/modi/rental/rental/infrastructure/client/dto/RentalItemInfoListResponse.java
@@ -6,7 +6,7 @@ import java.util.List;
 public record RentalItemInfoListResponse(
 		
 		List<RentalItemInfo> rentalItemInfoList,
-		Integer totalCount,
+		Long totalCount,
 		Integer totalPages
 ) {
 }
